+++ junos/provider.go
@@ -207,6 +207,7 @@ func Provider() *schema.Provider {
 			"junos_interface_logical":           dataSourceInterfaceLogical(),
 			"junos_interface_physical":          dataSourceInterfacePhysical(),
 			"junos_interfaces_physical_present": dataSourceInterfacesPhysicalPresent(),
+			"junos_routing_instance":            dataSourceRoutingInstance(),
 			"junos_system_information":          dataSourceSystemInformation(),
 		},
 		ConfigureContextFunc: configureProvider,