+++ junos/resource_routing_instance.go
@@ -24,7 +24,7 @@ type instanceOptions struct {
 	vtepSourceIf       string
 	instanceExport     []string
 	instanceImport     []string
-	interFace          []string
+	interFace          []string // to data_source
 	vrfExport          []string
 	vrfImport          []string
 }
@@ -137,11 +137,6 @@ func resourceRoutingInstance() *schema.Resource {
 
 					return
 				},
-			},
-			"interface": {
-				Type:     schema.TypeSet,
-				Computed: true,
-				Elem:     &schema.Schema{Type: schema.TypeString},
 			},
 		},
 	}
@@ -535,8 +530,5 @@ func fillRoutingInstanceData(d *schema.ResourceData, instanceOptions instanceOpt
 		if tfErr := d.Set("vrf_target_import", instanceOptions.vrfTargetImport); tfErr != nil {
 			panic(tfErr)
 		}
-	}
-	if tfErr := d.Set("interface", instanceOptions.interFace); tfErr != nil {
-		panic(tfErr)
 	}
 }