+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 
 ENHANCEMENTS:
 
-* resource/`junos_routing_instance`: add `interface` attribute to be able to read the list of interfaces in routing-instance
 * resource/`junos_security_zone`: add `interface` attribute to be able to read the list of interfaces in security-zone
+* add `junos_routing_instance` data source (to be able to read the list of interfaces in routing-instance)
 
 BUG FIXES: